+++ blaze/compatibility.py
@@ -3,6 +3,8 @@ from __future__ import absolute_import, division, print_function
 import sys
 from types import MethodType
 
+import pandas.util.testing as tm
+
 PY3 = sys.version_info[0] == 3
 PY2 = sys.version_info[0] == 2
 
@@ -103,3 +105,31 @@ try:
     from StringIO import StringIO
 except ImportError:
     from io import StringIO
+
+
+def assert_series_equal(left, right, check_names=True, **kwargs):
+    """Backwards compatibility wrapper for
+    ``pandas.util.testing.assert_series_equal``
+
+    Examples
+    --------
+    >>> import pandas as pd
+    >>> s = pd.Series(list('abc'), name='a')
+    >>> s2 = pd.Series(list('abc'), name='b')
+    >>> assert_series_equal(s, s2)  # doctest: +ELLIPSIS
+    Traceback (most recent call last):
+        ...
+    AssertionError: ...
+    >>> assert_series_equal(s, s2, check_names=False)
+
+    See Also
+    --------
+    pandas.util.testing.assert_series_equal
+    """
+    try:
+        return tm.assert_series_equal(left, right, check_names=check_names,
+                                      **kwargs)
+    except TypeError:
+        if check_names:
+            assert left.name == right.name
+        return tm.assert_series_equal(left, right, **kwargs)
